Got lerna and webpack dev server to work together
