--- addons.js
@@ -1,1 +1,0 @@
-module.exports = require("./dist/react-data-grid.ui-plugins");
