--- packages/react-data-grid-addons/main.js
@@ -1,1 +1,0 @@
-module.exports = require("./dist/react-data-grid-addons");
