+++ packages/react-data-grid-examples/index.js
@@ -1,0 +1,1 @@
+module.exports = require("./dist/");
