--- packages/react-data-grid-examples/main.js
@@ -1,1 +1,0 @@
-module.exports = require("./dist/examples");
