+++ packages/react-data-grid/package.json
@@ -2,7 +2,6 @@
   "name": "react-data-grid",
   "version": "2.0.0-dc-arch-revamp2285",
   "description": "The grid core",
-  "main": "main.js",
   "scripts": {
     "test": "npm test"
   },
