+++ packages/react-data-grid/src/index.js
@@ -1,11 +1,13 @@
 const Grid = require('./ReactDataGrid');
+import RowComparer from './RowComparer';
+import RowsContainer from './RowsContainer';
 
 module.exports = Grid;
 module.exports.Row = require('./Row');
 module.exports.Cell = require('./Cell');
 module.exports.HeaderCell = require('./HeaderCell');
-module.exports.RowComparer = require('./RowComparer');
-module.exports.RowsContainer = require('./RowsContainer');
+module.exports.RowComparer = RowComparer;
+module.exports.RowsContainer = RowsContainer;
 module.exports.editors = require('./editors');
 module.exports.utils = require('./utils');
 module.exports.shapes = require('./PropTypeShapes');
